requirements.txt
tqdm
pillow
numpy
albumentations